AT-CLI-006 I/O: Fix float32 rounding in SMV scaling formula
Problem Statement (spec-a-cli.md line 181):
- Spec requires: integer pixel = floor(min(65535, float*scale + adc))
- Test test_at_cli_006.py::test_explicit_scale_flag was failing with off-by-one error
- Pixel (5,5): expected 89, got 90

Root Cause:
- Float32 precision issue at value boundaries
- float_val=0.03999999538064003 (≈ 0.04) in float32
- float32: 0.04 * 1000 + 50 = 90.0 → floor = 90
- float64: 0.03999999538064003 * 1000 + 50 = 89.999... → floor = 89
- Test uses Python int() which operates in float64

Fix:
1. __main__.py line 1121: Convert intensity to float64 before scaling
   - Changed: scaled = intensity * scale + adc_offset
   - To: scaled = intensity.double() * scale + adc_offset

2. smv.py line 258: Convert image_data to float64 before scaling
   - Changed: scaled_data = image_data * scale + adc_offset
   - To: scaled_data = image_data.astype(np.float64) * scale + adc_offset

3. smv.py lines 259, 264, 269, 274: Use floor instead of round
   - Changed: np.round(clipped_data).astype(...)
   - To: np.floor(clipped_data).astype(...)
   - Rationale: Spec explicitly requires floor, not round

Verification:
- test_at_cli_006.py: 5/5 passed in 25.47s
- All AT-CLI-006 tests pass (autoscale, explicit scale, PGM)
- No regressions in smoke tests

Module: I/O
Tests: test_at_cli_006.py 5 passed

+++ src/nanobrag_torch/__main__.py
@@ -1116,7 +1116,9 @@ def main():
             roi_mask = intensity > threshold
 
             # Calculate scaled values
-            scaled = intensity * scale + adc_offset
+            # Per spec-a-cli.md line 181: integer pixel = floor(min(65535, float*scale + adc))
+            # Use float64 for scaling to match Python's int() behavior and avoid float32 rounding errors
+            scaled = intensity.double() * scale + adc_offset
             # Only apply scaling where intensity > 0 (inside ROI)
             scaled = torch.where(roi_mask, scaled, torch.zeros_like(scaled))
             # Clip to valid range and floor
